# Optimize `ULID.from_integer`