--- a/sig/ulid.rbs
+++ b/sig/ulid.rbs
@@ -87,7 +87,7 @@ class ULID
   def self.from_monotonic_generator: (MonotonicGenerator generator) -> ULID
   attr_reader milliseconds: Integer
   attr_reader entropy: Integer
-  def initialize: (milliseconds: Integer, entropy: Integer) -> void
+  def initialize: (milliseconds: Integer, entropy: Integer, ?integer: Integer) -> void
   def to_s: -> String
   def to_i: -> Integer
   alias hash to_i